+++ lib/taiko.js
@@ -142,6 +142,13 @@ const _closeBrowser = async () => {
   await closeConnection(promisesToBeResolvedBeforeCloseBrowser);
 };
 
+/**
+ * Gives CRI client object (a wrapper around Chrome DevTools Protocol). Refer https://github.com/cyrus-and/chrome-remote-interface
+ * This is useful while writing plugins or if use some API of [Chrome DevTools Protocol](https://chromedevtools.github.io/devtools-protocol/).
+ *
+ * @returns {Object}
+ */
+module.exports.client = () => getEventProxy(getClient());
 function getEventProxy(target) {
   if (!target) {
     return target;
@@ -163,14 +170,6 @@ function getEventProxy(target) {
   };
   return new Proxy(target, handler);
 }
-
-/**
- * Gives CRI client object (a wrapper around Chrome DevTools Protocol). Refer https://github.com/cyrus-and/chrome-remote-interface
- * This is useful while writing plugins or if use some API of [Chrome DevTools Protocol](https://chromedevtools.github.io/devtools-protocol/).
- *
- * @returns {Object}
- */
-module.exports.client = () => getEventProxy(getClient());
 
 /**
  * Allows switching between tabs and windows using URL or page title or Window name.
@@ -995,27 +994,9 @@ module.exports.dragAndDrop = async (source, destination) => {
 module.exports.hover = async (selector, options = {}) => {
   validate();
   options = setNavigationOptions(options);
-  const e = await findFirstElement(selector);
-  await scrollToElement(e);
-  if (defaultConfig.headful) {
-    await highlightElement(e);
-  }
-  const { x, y } = await domHandler.boundingBoxCenter(e.get());
-  const option = {
-    x: x,
-    y: y,
-  };
-  await doActionAwaitingNavigation(options, async () => {
-    Promise.resolve()
-      .then(() => {
-        option.type = 'mouseMoved';
-        return inputHandler.dispatchMouseEvent(option);
-      })
-      .catch((err) => {
-        throw new Error(err);
-      });
-  });
-  descEvent.emit('success', 'Hovered over the ' + description(selector, true));
+  const { hover } = require('./actions/hover');
+  const desc = await hover(selector, options);
+  descEvent.emit('success', desc);
 };
 
 /**
@@ -1030,6 +1011,7 @@ module.exports.hover = async (selector, options = {}) => {
  * @param {Object} options
  * @param {string[]} [options.waitForEvents = ['firstMeaningfulPaint']] - Events available to wait for ['DOMContentLoaded', 'loadEventFired', 'networkAlmostIdle', 'networkIdle', 'firstPaint', 'firstContentfulPaint', 'firstMeaningfulPaint']
  */
+const { focus } = require('./actions/focus');
 module.exports.focus = async (selector, options = {}) => {
   validate();
   options = setNavigationOptions(options);
@@ -1037,7 +1019,7 @@ module.exports.focus = async (selector, options = {}) => {
     if (defaultConfig.headful) {
       await highlightElement(await findFirstElement(selector));
     }
-    await _focus(selector);
+    await focus(selector);
   });
   descEvent.emit('success', 'Focussed on the ' + description(selector, true));
 };
@@ -1093,7 +1075,7 @@ module.exports.write = async (text, into, options = { delay: 0 }) => {
       for (let elem of elems) {
         try {
           if (!elem.focused) {
-            await _focus(elem);
+            await focus(elem);
           }
           activeElement = await runtimeHandler.activeElement();
           if (activeElement.notWritable) {
@@ -1176,7 +1158,7 @@ module.exports.clear = async (selector, options = {}) => {
       async () => {
         try {
           if (selector) {
-            await _focus(selector);
+            await focus(selector);
           }
         } catch (_) {}
         return !(await runtimeHandler.activeElement()).notWritable;
@@ -2759,32 +2741,6 @@ module.exports.getConfig = getConfig;
  * @param {string} [options.highlightOnAction = 'false' ] Option to highlight an element on action.
  */
 module.exports.setConfig = setConfig;
-
-const _focus = async (selector) => {
-  let elems = [selector];
-  if (isSelector(selector) || isElement(selector)) {
-    elems = await handleRelativeSearch(await findElements(selector), []);
-  }
-  let error;
-  for (const elem of elems) {
-    await scrollToElement(elem);
-    const result = await evaluate(elem, focusElement);
-    if (result.subtype != 'error') {
-      return;
-    } else if (result.subtype == 'error') {
-      error = result.description;
-    }
-  }
-  throw new Error(error);
-
-  function focusElement() {
-    if (this.disabled == true) {
-      throw new Error('Element is not focusable');
-    }
-    this.focus();
-    return true;
-  }
-};
 
 const promisesToBeResolvedBeforeCloseBrowser = [];
